Do not use ARROW_CHECK in production code. Consolidate python_to_arrow code
